libndt7-unix-openssl.h: allow to ovveride gettimeofday
If we define NDT7_GETTIMEOFDAY to point to another function, we can
then use, e.g., clock_gettime where available.

We still have the limitation that we're bound to struct timeval, yet
this does not seem to be a very bad deal for us, after all.

In fact, timeval has microsecond resolution, which is good enough
for our purposes (we probably need 10-ms precision here).